Detour GetModuleHandleExW function in x86
vcruntime dispatches tasks to Windows thread pool. When task finishes
the completion callback is invoked. In order to ensure the dll that
contais that callback code is still loaded, the refcount for the dll is
incremented (via GetModuleHandleExW) when task is scheduled, and
decremented (vie FreeLibrary) after callback finishes.

FreeLibrary called with a handle to unregistered module returns an
error, which is converted into unhandled exception and resulting in
crash.